+++ src/lcd/sainsmartks.cxx
@@ -29,9 +29,9 @@
 
 using namespace upm;
 
-SAINSMARTKS::SAINSMARTKS(uint8_t rs,  uint8_t enable,
-                         uint8_t d0, uint8_t d1, uint8_t d2, uint8_t d3,
-                         uint8_t keypad)
+SAINSMARTKS::SAINSMARTKS(int rs,  int enable,
+                         int d0, int d1, int d2, int d3,
+                         int keypad)
   : Lcm1602(rs, enable, d0, d1, d2, d3),
     m_aioKeypad(keypad)
 {
